verify biometry register request data (#135)
* verify biometry register request data

* fix error messages
