+++ Cargo.toml
@@ -43,6 +43,7 @@ tower_governor = "0.4"
 # UI embedding
 rust-embed = { version = "8.5", features = ["include-exclude"] }
 mime_guess = "2.0"
+base64 = "0.22.1"
 
 [build-dependencies]
 tonic-build = { version = "0.12" }
